backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
asyncpg==0.29.0
anthropic==0.39.0
qdrant-client==1.12.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.17
